+++ Tests/Fixtures/Block.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Sonata\PageBundle\Tests\Fixtures;
+
+use Sonata\PageBundle\Entity\BaseBlock;
+
+class Block extends BaseBlock
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getId()
+    {
+    }
+}